コンポーネント切り分けた
